# practice/src/main/java/ru/t1/practice/dto/TechnologyListResponse.java
package ru.t1.practice.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Data;

import java.util.List;

@Data
@AllArgsConstructor
@Schema(description = "Список технологий")
public class TechnologyListResponse {
    private List<TechnologyResponse> technologies;
}